+++ NEWS.md
@@ -1,5 +1,6 @@
 # shinybusy (unreleased)
 
+* Internal: use [{packer}](https://github.com/JohnCoene/packer) to manage JavaScript assets.
 * New function `update_modal_spinner()` to update modal spinner text ([#19](https://github.com/dreamRs/shinybusy/issues/19)).